apply patch from Johannes Lahti
Report:
I noticed that the playback gets stuck. After a few second it starts to
loop couple of frames endlessly. The reason seems to be that if the decoder
uses all the data that is given the program flow gets stuck in the while
loop.

Also It seems that if the decoder uses only a part of the data, the
remainder is never accepted by the decoder and it always returns 0. So
maybe the while loop is not necessary and we can just put the remainder
in the cache.