autotools: Major clean up of how autotools are used
This fixes the libtool issue with gst-sh-mobile. This is a wholesale
cleanup of the autotools stuff, largely based on TI’s plugins.

Also simplified the checks for uiomux, libshveu & libshcodecs as it
made the changes simpler and I think that we can rely on pkg_config and
the correct versions now.