SH buffers: Introduce a macro to check for SH peer.
Each SH plugin should be modified to check it's pads are connected
to another SH plugin. When this happens, the plugin can pass SH
buffers around.

Also, in most common cases the SH buffer can be directly mapped to
userspace such that non-SH plugins can access the data. There are
situations when this is not possible (e.g. alignment of chroma
plane), so plugins must allocate a normal gst buffer and copy the
data to this.